+++ src/main/resources/db/migration/V5__update_audit_trail.sql
@@ -1,0 +1,1 @@
+ALTER TABLE audit_trail ADD COLUMN user_message VARCHAR(2048);