+++ start.sh
@@ -11,7 +11,7 @@ help() {
   echo "-d, --no-daemon:    Do not run containers as daemons. This makes the logs visible in the console"
   echo "-l, --local-images: This will build the images for server and app locally instead of downloading images from GitHub."
   echo "                    To start it faster the next time, just start it without this option"
-  echo "-c, --cache:        Use Docker cache, default is to disable cache to ensure everything is rebuilt"
+  echo "-c, --cache:        Use Docker cache when building images locally. Ddefault is to disable cache to ensure everything is rebuilt"
   echo "-h, --help:         Show help"
   echo
   echo "If the startup fails, it might be because you have to little memory on your computer. Try to close other programs while you start up."
@@ -22,7 +22,7 @@ help() {
 DAEMON=-d
 COMPOSE_FILE=compose.yaml
 LOCAL_IMAGES=False
-NO_CACHE="--no-cache"
+NO_CACHE=True
 
 while [[ "$1" == "--"* || "$1" == "-"* ]]; do
   case $1 in
@@ -35,7 +35,7 @@ while [[ "$1" == "--"* || "$1" == "-"* ]]; do
       shift 1
       ;;
     --cache|-c)
-      NO_CACHE=""
+      NO_CACHE=False
       shift 1
       ;;
     --help|-h)
@@ -57,22 +57,33 @@ if [[ ! -f .env ]]; then
 fi
 
 if [[ "$LOCAL_IMAGES" == "True" ]]; then
-  echo "Building images locally..."
-  docker compose -f "$COMPOSE_FILE" build "$NO_CACHE" --progress plain
+  if [[ "$NO_CACHE" == "True" ]]; then
+  echo "Building images locally without cache..."
+    docker compose --progress plain -f "$COMPOSE_FILE" build --no-cache
+  else
+  echo "Building images locally with cache..."
+    docker compose --progress plain -f "$COMPOSE_FILE" build
+  fi
 fi
 
 echo "Starting Simple Invoice"
 echo
 
+source .env
+
+if [ -z "$APP_PORT" ]; then
+  APP_PORT=8000
+fi
+
 if [ -n "$DAEMON" ]; then
-  docker compose -f "$COMPOSE_FILE"  --progress plain up "$DAEMON"
+  docker compose --progress plain -f "$COMPOSE_FILE" up "$DAEMON"
   echo
-  echo "Simple Invoice App can be reached in a web browser at http://localhost:8000"
+  echo "Simple Invoice App can be reached in a web browser at http://localhost:$APP_PORT"
   echo "Run './stop.sh' to stop Simple Invoice"
 else
   echo
-  echo "Simple Invoice App can be reached in a web browser at http://localhost:8000"
+  echo "Simple Invoice App can be reached in a web browser at http://localhost:$APP_PORT"
   echo "Press Ctrl+C, and then run './stop.sh' to stop Simple Invoice"
   echo
-  docker compose -f "$COMPOSE_FILE" --progress plain up
+  docker compose --progress plain -f "$COMPOSE_FILE" up
 fi